# specification/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Swagger Microservices Bank
  description: Microservices Bank
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: apiteam@swagger.io
    url: http://swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://editor.swagger.io/api
paths:
  /clients/personal:
    get:
      tags:
        - CLIENT
      summary: Returns all clients from the system
      description: Returns all clients from the system
      operationId: findClients
      responses:
        '200':
          description: client response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      tags:
        - CLIENT
      summary: Creates a new client in the bank.  Duplicates are not allowed
      description: Creates a new client in the bank.  Duplicates are not allowed
      operationId: addClient
      requestBody:
        description: ClientDto to add to the bank
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewClientDto'
      responses:
        '200':
          description: client response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /clients/personal/{id}:
    get:
      tags:
        - CLIENT
      summary: Returns a user based on a id
      description: Returns a user based on a id
      parameters:
        - name: id
          in: path
          description: Id of client to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: client response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'


    put:
      tags:
        - CLIENT
      summary: Update client
      description: This can only be done by the logged in client.
      operationId: updateClient
      parameters:
        - name: id
          in: path
          description: id of client to fetch
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewClientDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewClientDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewClientDto'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        '400':
          description: Invalid identity number
        '404':
          description: Client not found
        '405':
          description: Validation exception
    delete:
      tags:
        - CLIENT
      summary: Deletes a single client
      description: Deletes a single client
      parameters:
        - name: id
          in: path
          description: Identity number of client to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: client deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /clients/business:
    get:
      tags:
        - CLIENT
      summary: Returns all clientsBusiness from the system
      description: Returns all clientsBusiness from the system
      operationId: findClientsBusiness
      responses:
        '200':
          description: clientBusiness response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientBusinessDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      tags:
        - CLIENT
      summary: Creates a new client business in the bank.  Duplicates are not allowed
      description: Creates a new client business in the bank.  Duplicates are not allowed
      operationId: addClientBusiness
      requestBody:
        description: ClientBusinessDto to add to the bank
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewClientBusinessDto'
      responses:
        '200':
          description: client response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientBusinessDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /clients/business/{id}:
    get:
      tags:
        - CLIENT
      summary: Returns a clientBusiness based on a id
      description: Returns a clientBusiness based on a id
      parameters:
        - name: id
          in: path
          description: Id of clientBusiness to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: clientBusiness response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientBusinessDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'


    put:
      tags:
        - CLIENT
      summary: Update clientBusiness
      description: This can only be done by the logged in clientBusiness.
      operationId: updateClientBusiness
      parameters:
        - name: id
          in: path
          description: id of clientBusiness to fetch
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent clientBusiness
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewClientBusinessDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewClientBusinessDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewClientBusinessDto'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientBusinessDto'
        '400':
          description: Invalid identity number
        '404':
          description: ClientBusiness not found
        '405':
          description: Validation exception
    delete:
      tags:
        - CLIENT
      summary: Deletes a single clientBusiness
      description: Deletes a single clientBusiness
      parameters:
        - name: id
          in: path
          description: Identity number of clientBusiness to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: clientBusiness deleted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
  /account:
    get:
      tags:
        - ACCOUNT
      summary: Returns all accounts from the system
      description: Returns all accounts from the system
      operationId: findAccounts
      responses:
        '200':
          description: account response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
  /account/banking:
    post:
      tags:
        - ACCOUNT
      summary: Creates a new account in the bank.  Duplicates are not allowed
      description: Creates a new account in the bank.  Duplicates are not allowed
      operationId: addAccount
      requestBody:
        description: Account to add to the bank
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAccountDto'
      responses:
        '200':
          description: account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /account/{accountId}/products:
    post:
      tags:
        - ACCOUNT
      summary: Associate a product to a customer account
      description: Associate a product to a customer account
      parameters:
        - name: accountId
          in: path
          description: Id of account to fetch
          required: true
          schema:
            type: string
      requestBody:
        description: Products to a customer account
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductListDto'
      responses:
        '200':
          description: account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
    ClientDto:
      allOf:
        - $ref: '#/components/schemas/NewClientDto'
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
    NewClientDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            name:
              type: "string"
            identityType:
              type: "string"
            identityNumber:
              type: "string"
            segmentType:
              type: "string"
            customerType:
              type: "string"
            address:
              type: "string"
            phoneNumber:
              type: "string"
    ClientBusinessDto:
      allOf:
        - $ref: '#/components/schemas/NewClientBusinessDto'
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
    NewClientBusinessDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            name:
              type: "string"
            identityType:
              type: "string"
            identityNumber:
              type: "string"
            segmentType:
              type: "string"
            customerType:
              type: "string"
            address:
              type: "string"
            phoneNumber:
              type: "string"
            holderList:
              type: "array"
              items:
                properties:
                  holderName:
                    type: "string"
            signatoryList:
              type: "array"
              items:
                properties:
                  signatoryName:
                    type: "string"
              minItems: 1
...
    AccountDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
            productList:
              type: "array"
              items:
                $ref: "#/components/schemas/ProductAccountDto"
            client:
              $ref: "#/components/schemas/ClientAccountDto"
    NewAccountDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            client:
              type: "object"
              properties:
                id:
                  type: "string"
            productList:
              type: "array"
              items:
                $ref: "#/components/schemas/NewProductAccountDto"
    NewProductListDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            productList:
              type: "array"
              items:
                $ref: "#/components/schemas/NewProductAccountDto"
    NewProductAccountDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
...
    ProductAccountDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
            name:
              type: "string"
            type:
              type: "string"
            balance:
              type: "number"
              format: "double"
    ClientAccountDto:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: "string"
            name:
              type: "string"
            customerType:
              type: "string"
...
    ErrorDto:
      type: object
      required:
        - message
      properties:
        message:
          type: string